Move CalmRecord and associated fixtures to common source_model
