+++ .sbt_metadata/calm_api_client.json
@@ -2,6 +2,7 @@
   "id" : "calm_api_client",
   "folder" : "calm_adapter/calm_api_client",
   "dependencyIds" : [
+    "source_model",
     "flows"
   ]
 }
