+++ calm_adapter/calm_adapter/src/main/scala/uk/ac/wellcome/calm_adapter/CalmAdapterWorkerService.scala
@@ -9,11 +9,12 @@ import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.MessageSender
 import uk.ac.wellcome.messaging.sns.NotificationMessage
 import uk.ac.wellcome.messaging.sqs.SQSStream
-import uk.ac.wellcome.platform.calm_api_client.{CalmQuery, CalmRecord}
+import uk.ac.wellcome.platform.calm_api_client.CalmQuery
 import uk.ac.wellcome.storage.Version
 import uk.ac.wellcome.storage.s3.S3ObjectLocation
 import uk.ac.wellcome.typesafe.Runnable
 import weco.catalogue.source_model.CalmSourcePayload
+import weco.catalogue.source_model.calm.CalmRecord
 import weco.flows.FlowOps
 
 import scala.concurrent.{ExecutionContext, Future}
