+++ calm_adapter/calm_adapter/src/main/scala/uk/ac/wellcome/calm_adapter/CalmStore.scala
@@ -1,7 +1,6 @@
 package uk.ac.wellcome.calm_adapter
 
 import grizzled.slf4j.Logging
-import uk.ac.wellcome.platform.calm_api_client.CalmRecord
 import uk.ac.wellcome.storage.s3.S3ObjectLocation
 import uk.ac.wellcome.storage.{
   Identified,
@@ -9,6 +8,7 @@ import uk.ac.wellcome.storage.{
   StorageError,
   Version
 }
+import weco.catalogue.source_model.calm.CalmRecord
 import weco.catalogue.source_model.store.SourceVHS
 
 class CalmStore(store: SourceVHS[CalmRecord]) extends Logging {
