+++ calm_adapter/calm_adapter/src/main/scala/uk/ac/wellcome/calm_adapter/Main.scala
@@ -9,10 +9,8 @@ import uk.ac.wellcome.typesafe.config.builders.AkkaBuilder
 import uk.ac.wellcome.typesafe.config.builders.EnrichConfig._
 import uk.ac.wellcome.messaging.typesafe.{SNSBuilder, SQSBuilder}
 import uk.ac.wellcome.json.JsonUtil._
-import uk.ac.wellcome.platform.calm_api_client.{
-  AkkaHttpCalmApiClient,
-  CalmRecord
-}
+import uk.ac.wellcome.platform.calm_api_client.AkkaHttpCalmApiClient
+import weco.catalogue.source_model.calm.CalmRecord
 import weco.catalogue.source_model.config.SourceVHSBuilder
 
 object Main extends WellcomeTypesafeApp {
