+++ calm_adapter/calm_api_client/src/main/scala/uk/ac/wellcome/platform/calm_api_client/CalmXmlRequest.scala
@@ -1,5 +1,7 @@
 package uk.ac.wellcome.platform.calm_api_client
 
+import weco.catalogue.source_model.calm.CalmRecord
+
 import scala.xml.Elem
 
 trait CalmXmlRequest {
