+++ calm_adapter/calm_api_client/src/test/scala/uk/ac/wellcome/platform/calm_api_client/CalmApiClientTest.scala
@@ -16,6 +16,7 @@ import uk.ac.wellcome.platform.calm_api_client.fixtures.{
   CalmApiClientFixtures,
   CalmResponseGenerators
 }
+import weco.catalogue.source_model.calm.CalmRecord
 
 import scala.concurrent.ExecutionContext.Implicits.global
