+++ calm_adapter/calm_api_client/src/test/scala/uk/ac/wellcome/platform/calm_api_client/CalmXmlResponseTest.scala
@@ -1,10 +1,11 @@
 package uk.ac.wellcome.platform.calm_api_client
 
 import java.time.Instant
+
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
-
 import akka.http.scaladsl.model.headers.{Cookie, HttpCookiePair}
+import weco.catalogue.source_model.calm.CalmRecord
 
 class CalmXmlResponseTest extends AnyFunSpec with Matchers {
