+++ common/source_model/src/test/scala/weco/catalogue/source_model/generators/CalmRecordGenerators.scala
@@ -1,11 +1,11 @@
-package uk.ac.wellcome.platform.transformer.calm.generators
+package weco.catalogue.source_model.generators
 
-import org.scalatest.funspec.AnyFunSpec
-import org.scalatest.matchers.should.Matchers
-import uk.ac.wellcome.models.work.generators.IdentifiersGenerators
-import uk.ac.wellcome.platform.transformer.calm.CalmRecord
+import uk.ac.wellcome.fixtures.RandomGenerators
+import weco.catalogue.source_model.calm.CalmRecord
 
-trait CalmRecordGenerators extends IdentifiersGenerators {
+trait CalmRecordGenerators extends RandomGenerators {
+
+  def createCalmRecordId: String = randomUUID.toString
 
   def createCalmRecordWith(fields: (String, String)*): CalmRecord = {
     // Roll up the fields into Map[String, List[String]]
@@ -24,28 +24,11 @@ trait CalmRecordGenerators extends IdentifiersGenerators {
     }
 
     CalmRecord(
-      id = createCalmRecordID,
+      id = createCalmRecordId,
       retrievedAt = randomInstant,
       data = data
     )
   }
 
   def createCalmRecord: CalmRecord = createCalmRecordWith()
-}
-
-class CalmRecordGeneratorsTest
-    extends AnyFunSpec
-    with Matchers
-    with CalmRecordGenerators {
-  it("assembles the data correctly") {
-    val record = createCalmRecordWith(
-      "Place" -> "London",
-      "Place" -> "Paris",
-      "Date" -> "2020"
-    )
-
-    record.data shouldBe Map(
-      "Place" -> List("London", "Paris"),
-      "Date" -> List("2020"))
-  }
 }
