--- pipeline/transformer/transformer_calm/src/main/scala/uk/ac/wellcome/platform/transformer/calm/CalmRecord.scala
@@ -1,10 +1,0 @@
-package uk.ac.wellcome.platform.transformer.calm
-
-import java.time.Instant
-
-case class CalmRecord(
-  id: String,
-  data: Map[String, List[String]],
-  retrievedAt: Instant,
-  published: Boolean = false
-)
