+++ pipeline/transformer/transformer_calm/src/main/scala/uk/ac/wellcome/platform/transformer/calm/transformers/CalmNotes.scala
@@ -1,13 +1,10 @@
 package uk.ac.wellcome.platform.transformer.calm.transformers
 
 import uk.ac.wellcome.models.work.internal._
-import uk.ac.wellcome.platform.transformer.calm.{
-  CalmOps,
-  CalmRecord,
-  NormaliseText
-}
+import uk.ac.wellcome.platform.transformer.calm.{CalmRecordOps, NormaliseText}
+import weco.catalogue.source_model.calm.CalmRecord
 
-object CalmNotes extends CalmOps {
+object CalmNotes extends CalmRecordOps {
   private val notesMapping = List(
     ("AdminHistory", BiographicalNote(_)),
     ("CustodHistory", OwnershipNote(_)),
