+++ pipeline/transformer/transformer_calm/src/test/scala/uk/ac/wellcome/platform/transformer/calm/services/CalmTransformerWorkerTest.scala
@@ -8,8 +8,6 @@ import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.sns.NotificationMessage
 import uk.ac.wellcome.models.work.internal.{IdentifierType, Work, WorkState}
 import uk.ac.wellcome.pipeline_storage.{PipelineStorageStream, Retriever}
-import uk.ac.wellcome.platform.transformer.calm.CalmRecord
-import uk.ac.wellcome.platform.transformer.calm.generators.CalmRecordGenerators
 import uk.ac.wellcome.storage.generators.S3ObjectLocationGenerators
 import uk.ac.wellcome.storage.s3.S3ObjectLocation
 import uk.ac.wellcome.storage.store.memory.MemoryTypedStore
@@ -18,8 +16,11 @@ import weco.catalogue.transformer.{
   TransformerWorker,
   TransformerWorkerTestCases
 }
-
 import java.util.UUID
+
+import weco.catalogue.source_model.calm.CalmRecord
+import weco.catalogue.source_model.generators.CalmRecordGenerators
+
 import scala.concurrent.ExecutionContext.Implicits.global
 
 class CalmTransformerWorkerTest
